pyproject.toml
[tool.poetry]
name = "ai-chess-experiments"
version = "0.1.0"
description = "Experiments with various AI approaches to play chess on Lichess"
authors = ["Duncan Haywood"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-chess = "^1.999"
berserk = "^0.13.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
numpy = "^1.26.0"  # For future ML experiments

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
